(PUP-8049) Ensure all acceptance tests are FIPS compatible.
Removed some and updated some tests to be able to run PA acceptance
on platforms in FIPS mode. This is being done get PA to run thru
rest of the CI pipe so any FIPS specific failures in other modules
can be identified. Separate followup work is needed to adjust the
failing set of tests.

Updated versions of acceptance tests to run in FIPS mode:
 - tests/puppet_apply_a_file_should_create_a_file_and_report_the_md5.rb
 - tests/ticket_1334_clientbucket_corrupted.rb
 - tests/ticket_6541_invalid_filebucket_files.rb
 - resource/file/content_attribute.rb

Skipping following (modules) tests:
 - tests/face/loadable_from_modules.rb
 - tests/modules/build/build_agent.rb